Use SetLoadFile(Text|Data)Callback instead of async fetches
We used to do emscripten_fetch for every asset we loaded, which meant we had
to effectively spin the MiniScript interpreter for however many frames it took
for the resource to load, and then we had to do a lot of the same things that
the raylib functions we were pretending to be did for us.

Now, we use raylib's built-in hooks, and some clever use of EM_ASYNC_JS, to
overwrite how LoadFileData and LoadFileText work. The fetches are still being
done asynchronously, but Emscripten hides that asynchronicity from our code
by suspending us until the file has been loaded (or until we know we can't load
the file). This allows us to use bog standard raylib functions like LoadFont as
if we were loading local files in a normal raylib project, and it all Just
Works(TM).